requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
python-decouple==3.6
requests==2.27.1
urllib3==1.26.8
Werkzeug==2.0.3
